xxw/oss-master/requirements.txt
-i https://pypi.douban.com/simple
amqp==2.3.2
asn1crypto==0.24.0
astroid==2.0.1
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
django-cors-headers==2.4.0
django-extensions==2.1.0
django-filter==2.0.0
django-requestlogging==1.0.1
django==2.0.7
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
markdown==2.6.11
mccabe==0.6.1
mysql-connector==2.1.6
mysqlclient==1.3.13
pycparser==2.18
pyjwt==1.6.4
pylint==2.0.1
pymysql==0.9.2
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
six==1.11.0
sqlalchemy==1.2.10
typed-ast==1.1.0; python_version < '3.7' and implementation_name == 'cpython'
urllib3==1.23
vine==1.1.4
wrapt==1.10.11
